hotfix: get assessment_db_dump working for conventional models
