+++ project/bmd/managers.py
@@ -1,0 +1,20 @@
+from utils.models import BaseManager
+
+
+class AssessmentSettingsManager(BaseManager):
+    assessment_relation = 'assessment'
+
+
+class LogicFieldManager(BaseManager):
+    assessment_relation = 'assessment'
+
+
+class SessionManager(BaseManager):
+    assessment_relation = 'endpoint__assessment'
+
+
+class ModelManager(BaseManager):
+    assessment_relation = 'session__endpoint__assessment'
+
+class SelectedModelManager(BaseManager):
+    assessment_relation = 'endpoint__assessment'
