+++ project/myuser/models.py
@@ -11,71 +11,10 @@ from django.template.loader import render_to_string
 from django.template import Context
 
 from utils.helper import SerializerHelper
-
-
-class HAWCMgr(BaseUserManager):
-    # from https://docs.djangoproject.com/en/1.5/topics/auth/customizing/
-    # also from UserManager(BaseUserManager)
-
-    def create_user(self, email, password=None, **extra_fields):
-        if not email:
-            raise ValueError('Users must have an email address')
-        if not password:
-            raise ValueError('Users must have an password')
-
-        now = timezone.now()
-        user = self.model(email=self.normalize_email(email),
-                          is_staff=False, is_active=True, is_superuser=False,
-                          last_login=now, date_joined=now, **extra_fields)
-
-        user.set_password(password)
-        user.save()
-        return user
-
-    def create_superuser(self, email, password=None, **extra_fields):
-        u = self.create_user(email, password, **extra_fields)
-        u.is_staff = True
-        u.is_active = True
-        u.is_superuser = True
-        u.save()
-        return u
-
-    def create_user_batch(self, email, first_name, last_name,
-                         pms=False, tms=False, rvs=False,
-                         welcome_email=False):
-        """
-        Create a HAWC user, assign to assessments, and optionally send a welcome-email.
-        Used for batch-creation of multiple users by administrators.
-
-        Assessment-are a list of assessment-ids or False if none are added.
-
-        Example method call:
-
-            create_hawc_user("hikingfan@gmail.com", "George", "Washington",
-                             pms=[1,2,3] tms=False, rvs=False,
-                             welcome_email=True)
-        """
-        user = self.create_user(
-            email=email,
-            password=self.make_random_password(),
-            first_name=first_name,
-            last_name=last_name)
-
-        if pms:
-            user.assessment_pms.add(*pms)
-
-        if tms:
-            user.assessment_teams.add(*tms)
-
-        if rvs:
-            user.assessment_reviewers.add(*rvs)
-
-        if welcome_email:
-            user.send_welcome_email()
-
+from . import managers
 
 class HAWCUser(AbstractBaseUser, PermissionsMixin):
-    objects = HAWCMgr()
+    objects = managers.HAWCMgr()
     email = models.EmailField(max_length=254, unique=True, db_index=True)
     first_name = models.CharField(_('first name'), max_length=30, blank=True)
     last_name = models.CharField(_('last name'), max_length=30, blank=True)
@@ -137,17 +76,10 @@ class HAWCUser(AbstractBaseUser, PermissionsMixin):
         msg.attach_alternative(html, "text/html")
         msg.send()
 
-    @classmethod
-    def assessment_qs(cls, assessment_id):
-        return cls.objects\
-            .filter(
-                models.Q(assessment_pms=assessment_id) |
-                models.Q(assessment_teams=assessment_id) |
-                models.Q(assessment_reviewers=assessment_id)
-            ).distinct()
-
 
 class UserProfile(models.Model):
+    objects = managers.UserProfileManager()
+    
     user = models.OneToOneField(HAWCUser, related_name='profile')
     HERO_access = models.BooleanField(default=False,
         verbose_name='HERO access',
