+++ project/study/models.py
@@ -349,6 +349,8 @@ class Study(Reference):
 
 
 class Attachment(models.Model):
+    objects = managers.AttachmentManager()
+    
     study = models.ForeignKey(Study, related_name="attachments")
     attachment = models.FileField(upload_to="study-attachment")
